refactor: update tooltip semantic color
