+++ .changeset/cold-seahorses-visit.md
@@ -1,0 +1,7 @@
+---
+'@consolelabs/tooltip': minor
+'@consolelabs/theme': minor
+'@consolelabs/core': minor
+---
+
+Update tooltip semantic color
