+++ packages/theme/src/components/tooltip.ts
@@ -1,58 +1,22 @@
-import { cva, VariantProps } from 'class-variance-authority'
 import clsx from 'clsx'
 
-const tooltipCva = cva(['rounded-lg px-3 py-2 text-xs font-semibold'], {
-  variants: {
-    theme: {
-      light: ['bg-white', 'text-neutral-800'],
-      dark: ['bg-neutral-800', 'text-white'],
-    },
-  },
-  defaultVariants: {
-    theme: 'light',
-  },
-})
+const tooltipClsx = ({ className = '' }: { className?: string }) =>
+  clsx(
+    'rounded-lg',
+    'px-3',
+    'py-2',
+    'text-xs',
+    'font-semibold',
+    'bg-background-tooltip',
+    'text-text-primary',
+    className,
+  )
 
-const tooltipArrowCva = cva([], {
-  variants: {
-    theme: {
-      light: ['fill-white'],
-      dark: ['fill-neutral-800'],
-    },
-  },
-  defaultVariants: {
-    theme: 'light',
-  },
-})
+const tooltipArrowClxs = 'fill-background-tooltip'
 
 const tooltipTriggerClsx = ({ className = '' }: { className?: string } = {}) =>
   clsx('w-fit', className)
 
-const tooltip = { tooltipCva, tooltipArrowCva, tooltipTriggerClsx }
-
-export const ARROW_OPTIONS = [
-  'none',
-  'top-start',
-  'top-center',
-  'top-end',
-  'right-start',
-  'right-center',
-  'right-end',
-  'bottom-start',
-  'bottom-center',
-  'bottom-end',
-  'left-start',
-  'left-center',
-  'left-end',
-] as const
-
-export type Arrow = (typeof ARROW_OPTIONS)[number]
-
-export type TooltipProps = VariantProps<typeof tooltipCva> & {
-  children: React.ReactNode
-  content: React.ReactNode
-  arrow?: Arrow | 'none'
-  className?: string
-}
+const tooltip = { tooltipClsx, tooltipArrowClxs, tooltipTriggerClsx }
 
 export { tooltip }
